fixed the git ignore to remove the java target folder because it isn't helpful

+++ CPP/Perception/CPP/Perception_Elements/Neural_Node.cpp
@@ -11,8 +11,8 @@ Perception::Layer::Neural_Node::Neural_Node()
 /// <summary>
 /// Instantiate a Neural Node and assigns it a value for input
 /// </summary>
-/// <param name="param_input">Short</param>
-Perception::Layer::Neural_Node::Neural_Node(short param_input)
+/// <param name="param_input">float</param>
+Perception::Layer::Neural_Node::Neural_Node(float param_input)
 {
 	this->input = param_input;
 }
@@ -20,9 +20,9 @@ Perception::Layer::Neural_Node::Neural_Node(short param_input)
 /// <summary>
 /// Instantiate a Neural Node and assigns it a value for input
 /// </summary>
-/// <param name="param_input">Short</param>
-/// <param name="param_weight">Short</param>
-Perception::Layer::Neural_Node::Neural_Node(short param_input, short param_weight)
+/// <param name="param_input">float</param>
+/// <param name="param_weight">float</param>
+Perception::Layer::Neural_Node::Neural_Node(float param_input, float param_weight)
 {
 	this->input = param_input;
 	this->weight = param_weight;
@@ -32,32 +32,32 @@ Perception::Layer::Neural_Node::Neural_Node(short param_input, short param_weigh
 /// <summary>
 /// Sets the input value of a Neural Node
 /// </summary>
-/// <param name="param_input">short</param>
-void Perception::Layer::Neural_Node::Set_Input(short param_input) {
+/// <param name="param_input">float</param>
+void Perception::Layer::Neural_Node::Set_Input(float param_input) {
 	this->input = param_input;
 }
 
 /// <summary>
 /// Gets the input value of the Neural Node
 /// </summary>
-/// <returns>short value of input from Neural Node</returns>
-short Perception::Layer::Neural_Node::Get_Input() {
+/// <returns>float value of input from Neural Node</returns>
+float Perception::Layer::Neural_Node::Get_Input() {
 	return this->input;
 }
 
 /// <summary>
 /// Sets the input value of a Neural Node
 /// </summary>
-/// <param name="param_weight">short</param>
-void Perception::Layer::Neural_Node::Set_Weight(short param_weight) {
+/// <param name="param_weight">float</param>
+void Perception::Layer::Neural_Node::Set_Weight(float param_weight) {
 	this->weight = param_weight;
 }
 
 /// <summary>
 /// Gets the Weight value of the Neural Node
 /// </summary>
-/// <returns>short value of weight from Neural Node</returns>
-short Perception::Layer::Neural_Node::Get_Weight() {
+/// <returns>float value of weight from Neural Node</returns>
+float Perception::Layer::Neural_Node::Get_Weight() {
 	return this->weight;
 }
